Improve memory usage in report initialization (BlueBrain/CoreNeuron#882)
* Use a vector for the report target gids in order to save memory
* Add intersection function for the report gids
* Clean up target vector after initialization
* sync gids and compartment id types to sonata

CoreNEURON Repo SHA: BlueBrain/CoreNeuron@47d8aef20c6e98d07ce57c07719bacb4f3c35e20

+++ src/coreneuron/io/reports/report_event.hpp
@@ -20,7 +20,7 @@ namespace coreneuron {
 
 #if defined(ENABLE_BIN_REPORTS) || defined(ENABLE_SONATA_REPORTS)
 struct VarWithMapping {
-    int id;
+    uint32_t id;
     double* var_value;
     VarWithMapping(int id_, double* v_)
         : id(id_)
@@ -28,7 +28,7 @@ struct VarWithMapping {
 };
 
 // mapping the set of variables pointers to report to its gid
-using VarsToReport = std::unordered_map<int, std::vector<VarWithMapping>>;
+using VarsToReport = std::unordered_map<uint64_t, std::vector<VarWithMapping>>;
 
 class ReportEvent: public DiscreteEvent {
   public:
